Adding HllSizeUtils
